+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 0.13.3 (tbd)
+## 0.13.3 (2025-02-18)
 
 - Implement `PartialSmt` (#372, #381).
 - Fix panic in `PartialMmr::untrack` (#382).